Cleanup of pickling code
